Refactoring and working on font atlas

+++ Chess/Chess/src/maths/calc.cpp
@@ -6,7 +6,7 @@ namespace Chess
 	{
 		float Calc::to_radians(float degrees)
 		{
-			return degrees * M_PI / 180.0f;
+			return (float)(degrees * M_PI / 180.0f);
 		}
 	}
 }
